added SetvariablesFunctions.C
